requirements.txt
flask==3.0.3
rembg==2.0.59
pillow==10.4.0
gunicorn==22.0.0
